*: implement rule to squash chainable tables into a single one
Fixes #197

This commit introduces the new rule to squash inner joins between tables that are chainable. This rule is yet experimental and thus can be used by using the `UNSTABLE_SQUASH_ENABLE` environment variable until it is proven to be stable and the results match the ones retrieved without this optimisation.

What is a join between chainable tables? That's an inner join between two tables that can be joined using some logic in which one table records are retrieved from some data in the other table rows because there is some sort of relationship between them. To assert such logic can be applied, the join needs to have certain filters in its condition.

- `repositories` is chainable with `remotes` iff `repositories.id = remotes.repository_id`
- `repositories` is chainable with `refs` iff `repositories.id = refs.repository_id`
- `remotes` is chainable with `refs` iff `refs.repository_id = remotes.repository_id`
- `refs` is chainable with `commits` iff `refs.hash = commits.hash`, `history_idx(refs.hash, commits.hash) >= 0` or `0 <= history_idx(refs.hash, commits.hash)`
- `refs` is chainable with `tree_entries` iff `commit_has_tree(refs.hash, tree_entries.tree_hash)`
- `refs` is chainable with `blobs` iff `commit_has_blob(refs.hash, blob.hash)`
- `commits` is chainable with `tree_entries` iff `commit_has_tree(commits.hash, tree_entries.tree_hash)` or `commits.tree_hash = tree_entries.tree_hash`
- `commits` is chainable with `blobs` iff `commit_has_blob(commits.hash, blob.hash)`
- `tree_entries` is chainable with `blobs` iff `tree_entries.entry_hash = blobs.hash`

Since chains are built hierarchically using the table hierarchy (repositories -> remotes -> refs -> commits -> tree_entries -> blobs), any chain can be joined with any table iff the last part of the chain and that table match the conditions stated before.

To allow this, there are several iterators, some for the same table, depending on the position in the chain that table is, that are composed in the analyzer rule.

Even if the chain is built hierarchically and thus it has a different schema ordering than the original schema, a mapping between the chain schema and the original one is computed so that the result of the chain is exactly the same as the one without the optimisation.

Notes about the experimental status of this feature:

- Lacks some integration tests that will be done on subsequent PRs.
- Tree entries table is currently under examination and depending on what we decide, the results it returns right now might need to change.
- Probably better to test it thoroughly and do some performance measurements before making this the default behaviour.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>