*: iterator should visit trees recursively after main tree
There was an inconsistency between the results returned by the iterators and the ones using the tables and joins because the tables look for all trees and visit them. Instead, the iterator chained with commits only visited the main tree. Now, there is another iterator for this case that visits the main tree and puts other trees that are found in a queue to visit them as well. This mimics the behaviour of joinin commits and tree entries.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>