*: don't read blob content if it's not used
Blob contents should not be read unless the blobs.content column is specifically used in the query (either in direct use or in a projection) to save memory. Operations such as aggregations may use a huge amount of memory without this optimisation.

Thanks to the column pushdown we are able to know if the column is in use, so if it is, just flag the iterators to let them know they should read the content.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>