gitbase: add support for siva files
Add support in RepositoryPool for handling siva files as repositories.

Added a new `AddSivaDir` method to `RepositoryPool`, which adds all the siva files inside that directory (a siva file is considered to be any file whose extension is `.siva`) and the direct child directories to allow bucketing.

Also some minor changes related to the RepositoryPool:
- NewRepositoryFromPath returns now *Repository for convenience.
- NewRepositoryPool returns now *RepositoryPool, as it always needs to be used with a pointer.
- Fixed golint warning about returning an unexported type in an exported method, which may be annoying by replacing the type with `sql.RowIter`.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>