gitbase: support for unpacked objects, integration tests added
This PR adds support for unpacked objects, which made the index contain different objects and not match the result of the query without indexes.
A bunch of integration tests for indexes of all tables testing they output the exact same thing as the same query without indexes have been added as well.

Because it's super important to ensure the index is closed so it can be released, each index iterator has now a call to `closeIndexOnError`, which closes the index if there is an error or EOF in an iterator, ensure it's *always* closed.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>