Do not use full path in plain repository IDs
The base path that is stripped to create the ID found as follows:

* If the path does not contain glob chars all of it
* If it contains glob pattern all path components previous to the first
pattern occurrence

This way we can add bucketed directories and have the repositories at
the same level. For example, adding "bucket/*":

    bucket/a/aa.git -> aa.git
    bucket/a/ab.git -> ab.git
    bucket/b/ba.git -> ba.git
    bucket/b/bb.git -> bb.git

Now a repository is only added once and an error log is printed if a new
repo with the same ID tries to be registered. This allows adding
complementary repo. Example, add new and old download dirs.

Signed-off-by: Javi Fontan <jfontan@gmail.com>