gitbase: close siva FS after use
When using siva filesystems the FS itself was closed by the garbage
collector. This caused to reach opened files limit when the tables did not
trigger garbage collection for some time. One of these cases is getting
remotes:

    select * from remotes;

Now when opening a siva FS a function is added to the repository to
close it.

Also added tests to check that the files are indeed closed. repository
and go-billy-siva are mocked to keep track of opened files.

Signed-off-by: Javi Fontan <jfontan@gmail.com>