dockerfile: include zero-config mysql client
Fixes #733

With this change, users won't even have to install the MySQL client
on their machines or care about setting the parameters every single
time they use it. By just `docker exec -it $IMAGE_NAME mysql` it
will connect to gitbase without requiring any kind of configuration.

Connection details are stored in `$HOME/.my.cnf`, which is generated
at the start of the gitbase server. Because gitbase user and password
can change when you run it, not only when you build the image, this
needs to be done on startup. That's why `init.sh` has been included,
which just creates the configuration for MySQL client and starts the
gitbase server.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>