command: skip a directory if gitbase has no permission to read it
Fixes #718

When a child of one of the paths given to the gitbase server is not
readable because of permissions, the repo finding process just
stopped and no repository was added to the pool.
This skips a directory if there are no permissions to read it, so
the process can continue adding repositories to the pool as usual.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>