skip references pointing to objects that are not commits
Fixes #745

On all iterators where references are used only hash references
are used. But we're not taking into account that hash references
can point to any kind of object, not only commit objects, which
are the only objects we care about for the references table.

This fix was already applied on the new commits iterator that
walks the history skipping unreachable commits, but it was not
consistent with the rest of the iterators, which only check if the
reference was a hash reference or not. Now this behavior is
consistent in all iterators.

`isIgnoredReference` function has been added to abstract all this
logic away so if we ever need to change the logic of this we will
only have to replace it in one place instead of being scattered
all over the codebase.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>