rule: parallelize projects calling UAST not under exchanges
Fixes #766

Since rows are iterated serially (except for exchange nodes), when
a Project node calls UAST, it's processed one at a time, which
underutilizes the resources in the machine.

When a Project calls UAST or UAST_MODE and is not under an Exchange
node, which already parallelizes its children, replaces the project
with a special node called parallelProject, which is essentially a
project that keeps up to N goroutines processing rows, where N is
the parallelism value of gitbase.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>